refactor(dashboard): promote ABI and param-shape utilities to shared
