+++ apps/dashboard/features/create-proposal/utils/argTree.ts
@@ -5,50 +5,19 @@ import {
   type Hex,
 } from "viem";
 
+import {
+  expectedLength,
+  parseArrayType,
+  shapeOf,
+  type ParamShape,
+} from "@/shared/utils/paramShape";
+
+export { expectedLength, parseArrayType, shapeOf, type ParamShape };
+
 export type ArgValue = string | ArgValue[];
-
-type ArrayInfo = { elementType: string; length: number | null };
-
-export const parseArrayType = (type: string): ArrayInfo | null => {
-  const match = type.match(/^(.*)\[(\d*)\]$/);
-  if (!match) return null;
-  return { elementType: match[1], length: match[2] ? Number(match[2]) : null };
-};
-
-const getComponents = (param: AbiParameter): readonly AbiParameter[] =>
-  (param as { components?: readonly AbiParameter[] }).components ?? [];
-
-/** What a parameter is, decided once. Every walk over an `AbiParameter` starts
- *  with the same array/tuple/leaf question; that preamble was copied into ten
- *  functions across four files, and the copies drifted. */
-export type ParamShape =
-  | { kind: "array"; element: AbiParameter; length: number | null }
-  | { kind: "tuple"; components: readonly AbiParameter[] }
-  | { kind: "leaf"; type: string };
-
-export const shapeOf = (param: AbiParameter): ParamShape => {
-  const array = parseArrayType(param.type);
-  if (array) {
-    return {
-      kind: "array",
-      element: { ...param, type: array.elementType } as AbiParameter,
-      length: array.length,
-    };
-  }
-  if (param.type === "tuple") {
-    return { kind: "tuple", components: getComponents(param) };
-  }
-  return { kind: "leaf", type: param.type };
-};
 
 const isComposite = (type: string): boolean =>
   parseArrayType(type) !== null || type === "tuple" || type.startsWith("tuple");
-
-export const expectedLength = (shape: ParamShape): number | null => {
-  if (shape.kind === "tuple") return shape.components.length;
-  if (shape.kind === "array") return shape.length;
-  return null;
-};
 
 /** The one place a container's declared size is compared with what it holds. Both
  *  halves matter: an extra entry is dropped on the way to the encoder, and a
