+++ apps/dashboard/shared/services/decoder/abi/etherscan.test.ts
@@ -1,8 +1,8 @@
-import { fetchAbi } from "@/features/create-proposal/utils/fetchAbi";
+import { fetchVerifiedAbi } from "@/shared/services/decoder/abi/etherscan";
 
 const VALID_ADDR = `0x${"a".repeat(40)}`;
 
-describe("fetchAbi", () => {
+describe("fetchVerifiedAbi", () => {
   const mockFetch = jest.fn();
 
   beforeEach(() => {
@@ -11,7 +11,7 @@ describe("fetchAbi", () => {
   });
 
   test("returns null for an invalid address", async () => {
-    await expect(fetchAbi(1, "not-an-address")).resolves.toBeNull();
+    await expect(fetchVerifiedAbi(1, "not-an-address")).resolves.toBeNull();
     expect(mockFetch).not.toHaveBeenCalled();
   });
 
@@ -27,7 +27,7 @@ describe("fetchAbi", () => {
         }),
     });
 
-    await expect(fetchAbi(1, VALID_ADDR)).resolves.toEqual(abi);
+    await expect(fetchVerifiedAbi(1, VALID_ADDR)).resolves.toEqual(abi);
     expect(mockFetch).toHaveBeenCalledTimes(1);
     const [calledUrl] = mockFetch.mock.calls[0] as [string];
     expect(calledUrl).toContain("/api/etherscan");
@@ -46,12 +46,12 @@ describe("fetchAbi", () => {
           result: "Contract source code not verified",
         }),
     });
-    await expect(fetchAbi(1, VALID_ADDR)).resolves.toBeNull();
+    await expect(fetchVerifiedAbi(1, VALID_ADDR)).resolves.toBeNull();
   });
 
   test("returns null on non-OK HTTP response", async () => {
     mockFetch.mockResolvedValue({ ok: false, json: () => Promise.resolve({}) });
-    await expect(fetchAbi(1, VALID_ADDR)).resolves.toBeNull();
+    await expect(fetchVerifiedAbi(1, VALID_ADDR)).resolves.toBeNull();
   });
 
   test("returns null when the result payload is not valid JSON", async () => {
@@ -60,6 +60,6 @@ describe("fetchAbi", () => {
       json: () =>
         Promise.resolve({ status: "1", message: "OK", result: "not-json" }),
     });
-    await expect(fetchAbi(1, VALID_ADDR)).resolves.toBeNull();
+    await expect(fetchVerifiedAbi(1, VALID_ADDR)).resolves.toBeNull();
   });
 });
